+++ src/Products/MailHost/interfaces.py
@@ -20,6 +20,7 @@ from zope.interface import Interface
 
 class IMailHost(Interface):
 
-    def send(messageText, mto=None, mfrom=None, subject=None, encode=None):
+    def send(messageText, mto=None, mfrom=None, subject=None, encode=None,
+             charset=None, msg_type=None):
         """Send mail.
         """